Renamed Service to PlayerService within the Players api
